// packages/beecli/sdk/beetrade-v2/models/CoreAlertTransportHttpCreateAlertRequest.ts
export interface CoreAlertTransportHttpCreateAlertRequest {
    /**
     * 
     * @type {BrokerAction}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    brokerAction?: BrokerAction;
    /**
     * 
     * @type {number}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    brokerageId: number;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    conditionSummary?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    data: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    expiresAt?: string;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    strategyId?: number;
    /**
     * 
     * @type {Array<NotificationType>}
     * @memberof CoreAlertTransportHttpCreateAlertRequest
     */
    types: Array<NotificationType>;
}