// packages/beecli/sdk/beetrade-v2/models/CoreAlertTransportHttpUpdateAlertRequest.ts
export interface CoreAlertTransportHttpUpdateAlertRequest {
    /**
     * 
     * @type {BrokerAction}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    brokerAction?: BrokerAction;
    /**
     * 
     * @type {number}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    brokerageId?: number;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    conditionSummary?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    data?: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    expiresAt?: string;
    /**
     * 
     * @type {string}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    strategyId?: number;
    /**
     * 
     * @type {Array<NotificationType>}
     * @memberof CoreAlertTransportHttpUpdateAlertRequest
     */
    types?: Array<NotificationType>;
}