// packages/beecli/sdk/beetrade-v2/models/CoreAlertTransportHttpUpdateAlertStatusRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CoreAlertTransportHttpUpdateAlertStatusRequest
 */
export interface CoreAlertTransportHttpUpdateAlertStatusRequest {
    /**
     * 
     * @type {boolean}
     * @memberof CoreAlertTransportHttpUpdateAlertStatusRequest
     */
    status?: boolean;
}

/**
 * Check if a given object implements the CoreAlertTransportHttpUpdateAlertStatusRequest interface.
 */
export function instanceOfCoreAlertTransportHttpUpdateAlertStatusRequest(value: object): value is CoreAlertTransportHttpUpdateAlertStatusRequest {
    return true;
}

export function CoreAlertTransportHttpUpdateAlertStatusRequestFromJSON(json: any): CoreAlertTransportHttpUpdateAlertStatusRequest {
    return CoreAlertTransportHttpUpdateAlertStatusRequestFromJSONTyped(json, false);
}

export function CoreAlertTransportHttpUpdateAlertStatusRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoreAlertTransportHttpUpdateAlertStatusRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function CoreAlertTransportHttpUpdateAlertStatusRequestToJSON(json: any): CoreAlertTransportHttpUpdateAlertStatusRequest {
    return CoreAlertTransportHttpUpdateAlertStatusRequestToJSONTyped(json, false);
}

export function CoreAlertTransportHttpUpdateAlertStatusRequestToJSONTyped(value?: CoreAlertTransportHttpUpdateAlertStatusRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
    };
}